:hammer: refactor: Refactor dir structure
