+++ frontend/src/common/Avatar/Avatar.stories.tsx
