+++ frontend/src/common/Avatar/Avatar.test.tsx
