+++ frontend/src/common/Avatar/Avatar.tsx
