+++ frontend/src/common/Avatar/index.ts
