+++ frontend/src/common/Box/Box.stories.ts
