+++ frontend/src/layout/Grid/Grid.tsx
@@ -55,6 +55,7 @@ const StyledGrid = styled.div<GridProps>`
       ${props.size.overflow && `overflow: ${props.size.overflow};`}
       ${props.size.overflowX && `overflow-x: ${props.size.overflowX};`}
       ${props.size.overflowY && `overflow-y: ${props.size.overflowY};`}
+      ${props.size.padding && `padding: ${props.theme.padding[props.size.padding]};`}
     `}
   ${(props) =>
     props.position &&
