+++ frontend/src/layout/Header/Header.tsx
@@ -1,11 +1,12 @@
 import styled from 'styled-components';
 import { ReactComponent as Logo } from '@/svgs/logo-sm.svg';
+import { Grid } from '../Grid';
 
 export const Header = () => {
   return (
-    <StyledHeader>
+    <Grid as="header" height="12rem" container="flex" justifyContent="space-between" alignItems="center">
       <Logo />
-    </StyledHeader>
+    </Grid>
   );
 };
