+++ frontend/tsconfig.json
@@ -24,7 +24,9 @@
       "@/svgs/*": ["src/assets/svgs/*"],
       "@/assets/*": ["src/assets/*"],
       "@/test/*": ["src/test/*"],
-      "@/components/*": ["src/components/*"],
+      "@/common/*": ["src/common/*"],
+      "@/error/*": ["src/error/*"],
+      "@/layout/*": ["src/layout/*"],
       "@/pages/*": ["src/pages/*"],
       "@/hooks/*": ["src/hooks/*"],
       "@/utils/*": ["src/utils/*"],
