+++ frontend/vite.config.ts
@@ -17,7 +17,9 @@ export default defineConfig({
       '@/svgs': path.resolve(__dirname, './src/assets/svgs'),
       '@/assets': path.resolve(__dirname, './src/assets'),
       '@/test': path.resolve(__dirname, './src/test'),
-      '@/components': path.resolve(__dirname, './src/components'),
+      '@/common': path.resolve(__dirname, './src/common'),
+      '@/error': path.resolve(__dirname, './src/error'),
+      '@/layout': path.resolve(__dirname, './src/layout'),
       '@/pages': path.resolve(__dirname, './src/pages'),
       '@/hooks': path.resolve(__dirname, './src/hooks'),
       '@/utils': path.resolve(__dirname, './src/utils'),
